changed code

+++ Code/DC2019_MASTER/DC2019_MASTER.ino
@@ -38,7 +38,7 @@ String readBTline(){
 
 void distance_detect(){
   long duration = 0;
-  int distance;
+  long distance;
   for (int i = 0; i<3; i++){
     // Clears the trigPin
     digitalWrite(trigPin, LOW);
